requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-RESTful==0.3.10
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Werkzeug==2.3.7
SQLAlchemy==2.0.21
WTForms==3.0.1
PyMySQL==1.1.0
PyJWT==2.8.0
requests==2.31.0
Pillow==10.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
gevent==23.9.1
bcrypt==4.0.1
email-validator==2.0.0 